requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
slack_sdk==3.33.1
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
yarl==1.13.1
